requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
google-generativeai
pydantic
